shard-trainer/pyproject.toml
[project]
name = "shard-trainer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.6",
    "pandas>=2.3.0",
    "python-chess>=1.999",
    "torch>=2.7.1",
    "tqdm>=4.67.1",
]
